fix terminate import that lead to import error (circular import)
