+++ yaetos/deploy_utils.py
@@ -1,0 +1,13 @@
+from yaetos.logger import setup_logging
+logger = setup_logging('Deploy')
+
+
+def terminate(error_message=None):
+    """
+    Method to exit the Python script. It will log the given message and then exit().
+    :param error_message:
+    """
+    if error_message:
+        logger.error(error_message)
+    logger.critical('The script is now terminating')
+    exit()
